fix: dedup eip-3009 test files [#12]
