+++ test/DebtTokenEIP3009Test.js
@@ -9,6 +9,7 @@ const { pack } = require("@ethersproject/solidity");
 const { hexlify } = require("@ethersproject/bytes");
 const { randomBytes } = require("@ethersproject/random");
 const { ecsign } = require("ethereumjs-util");
+const { ethers, upgrades } = require("hardhat");
 
 const th = testHelpers.TestHelper;
 const { toBN, assertRevert, dec, ZERO_ADDRESS, GAS_COMPENSATION, MIN_NET_DEBT } = th;
@@ -163,6 +164,85 @@ contract("DebtToken - EIP-3009 Functionality", async () => {
     await hre.network.provider.send("hardhat_reset");
     signers = await ethers.getSigners();
     [owner, alice, bob, carol, dennis] = await signers.splice(0, 5);
+  });
+
+  // Simple verification tests from DebtTokenEIP3009SimpleTest.js
+  describe("Simple EIP-3009 Verification", () => {
+    it("should have EIP-3009 functions available in the contract", async () => {
+      // Deploy DebtToken directly to test the implementation
+      const DebtToken = await ethers.getContractFactory("DebtToken");
+
+      // Deploy with proxy
+      const debtToken = await upgrades.deployProxy(
+        DebtToken,
+        [
+          owner.address, // troveManager
+          owner.address, // stabilityPool
+          owner.address, // borrowerOperations
+        ],
+        {
+          unsafeAllow: ["constructor", "state-variable-immutable"],
+        },
+      );
+
+      await debtToken.deployed();
+
+      // Verify contract has the expected functions
+      assert.isNotNull(debtToken.transferWithAuthorization);
+      assert.isNotNull(debtToken.receiveWithAuthorization);
+      assert.isNotNull(debtToken.cancelAuthorization);
+      assert.isNotNull(debtToken.authorizationState);
+
+      // Test authorizationState function
+      const testNonce = ethers.utils.hexZeroPad("0x1", 32);
+      const state = await debtToken.authorizationState(alice.address, testNonce);
+      assert.isFalse(state);
+    });
+
+    it("should maintain existing ERC20 functionality", async () => {
+      // Deploy a fresh instance
+      const DebtToken = await ethers.getContractFactory("DebtToken");
+      const debtToken = await upgrades.deployProxy(
+        DebtToken,
+        [owner.address, owner.address, owner.address],
+        {
+          unsafeAllow: ["constructor", "state-variable-immutable"],
+        },
+      );
+
+      // Verify basic ERC20 functions still work
+      const name = await debtToken.name();
+      assert.equal(name, "USD for Filecoin Community");
+
+      const symbol = await debtToken.symbol();
+      assert.equal(symbol, "USDFC");
+
+      const decimals = await debtToken.decimals();
+      assert.equal(decimals, 18);
+
+      const totalSupply = await debtToken.totalSupply();
+      assert.equal(totalSupply.toString(), "0");
+    });
+
+    it("should have correct EIP-3009 type hashes", async () => {
+      // Check that the implementation includes the correct constants
+      const DebtToken = await ethers.getContractFactory("DebtToken");
+      const bytecode = DebtToken.bytecode;
+
+      // EIP-3009 type hashes as hex strings (without 0x prefix)
+      const transferAuthTypeHash =
+        "7c7c6cdb67a18743f49ec6fa9b35f50d52ed05cbed4cc592e13b44501c1a2267";
+      const receiveAuthTypeHash =
+        "d099cc98ef71107a616c4f0f941f04c322d8e254fe26b3c6668db87aae413de8";
+      const cancelAuthTypeHash = "158b0a9edf7a828aad02f63cd515c68ef2f50ba807396f6d12842833a1597429";
+
+      // Basic check that these constants appear in the bytecode
+      const hasTransferAuth = bytecode.includes(transferAuthTypeHash);
+      const hasReceiveAuth = bytecode.includes(receiveAuthTypeHash);
+      const hasCancelAuth = bytecode.includes(cancelAuthTypeHash);
+
+      assert.isTrue(hasTransferAuth || hasReceiveAuth || hasCancelAuth);
+    });
   });
 
   const testCorpus = ({ withProxy = false }) => {
@@ -683,7 +763,7 @@ contract("DebtToken - EIP-3009 Functionality", async () => {
             hexlify(r),
             hexlify(s),
           ),
-        "DebtToken: caller must be the payee",
+        "DebtToken: caller must be the recipient",
       );
     });
